fix: properly highlight selected item
