+++ apps/website/src/app/docs/packages/[package]/[version]/[item]/layout.tsx
@@ -2,7 +2,7 @@ import type { PropsWithChildren } from 'react';
 import { Providers } from './providers';
 import { CmdKDialog } from '~/components/CmdK';
 
-export default function SidebarLayout({ children }: PropsWithChildren) {
+export default function ItemLayout({ children }: PropsWithChildren) {
 	return (
 		<Providers>
 			<>{children}</>
